new page for statically loaded listings
