+++ sites/public/pages/eligibility/income.tsx
@@ -61,7 +61,7 @@ const EligibilityIncome = () => {
       params.seniorHousing = false
     }
 
-    return `/listings?${encodeToFrontendFilterString(params)}`
+    return `/listings/filtered?${encodeToFrontendFilterString(params)}`
   }
 
   return (
